FASE2-028: Incluye en el detalle de la transacción los puntos extras dependiendo de cada suscripción (#69)
- Permite visualizar puntos extra en cada transacción.
- Permite visualizar los puntos de cada transacción en el datatable de transacciones.
- Permite redondear el total del monto de las transacciones.

+++ petromiles-backend/src/entities/user-points.entity.ts
@@ -35,11 +35,11 @@ import { StateName } from '@/enums/state.enum';
       .groupBy('userClient.email'),
 })
 export class ClientPoints {
-  @Transform(dollars => parseInt(dollars).toFixed(2))
+  @Transform(dollars => parseFloat(dollars).toFixed(2))
   @ViewColumn()
   dollars: number;
 
-  @Transform(points => parseInt(points).toFixed(2))
+  @Transform(points => parseFloat(points).toFixed(2))
   @ViewColumn()
   points: number;
